Moved setUpActions helpers to a separate file
